+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/AVM2LocalData.java
@@ -118,6 +118,8 @@ public class AVM2LocalData extends BaseLocalData {
 
     public boolean inGetLoops = false;
 
+    public Set<Integer> seenMethods = new HashSet<>();
+
     public AVM2LocalData() {
 
     }
@@ -165,6 +167,7 @@ public class AVM2LocalData extends BaseLocalData {
         finallyTargetParts = localData.finallyTargetParts;
         pushDefaultPart = localData.pushDefaultPart;
         finallyKinds = localData.finallyKinds;
+        seenMethods = localData.seenMethods;
     }
 
     public AVM2ConstantPool getConstants() {