+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - #1114 Script search results dialogs closing on swf close
 - #1159 Regexp syntax hilight when not a regexp (only division)
+- #1227 AS3 avoid recursion (stackoverflow) caused by newfunction instruction
 
 ## [13.0.2] - 2021-02-10
 ### Changed